--- foundation_model/stable_diffusion/generate_utils.py
@@ -1,26 +1,0 @@
-from diffusers import DiffusionPipeline
-from diffusers.loaders import LoraLoaderMixin
-import torch
-
-
-def load_lora_weights(unet, text_encoder, input_dir):
-    lora_state_dict, network_alphas = LoraLoaderMixin.lora_state_dict(input_dir)
-    LoraLoaderMixin.load_lora_into_unet(
-        lora_state_dict, network_alphas=network_alphas, unet=unet
-    )
-    LoraLoaderMixin.load_lora_into_text_encoder(
-        lora_state_dict, network_alphas=network_alphas, text_encoder=text_encoder
-    )
-    return unet, text_encoder
-
-
-def get_pipeline(model_dir, lora_weights_dir=None):
-    pipeline = DiffusionPipeline.from_pretrained(model_dir, torch_dtype=torch.float16)
-    if lora_weights_dir:
-        unet = pipeline.unet
-        text_encoder = pipeline.text_encoder
-        print(f"Loading LoRA weights from {lora_weights_dir}")
-        unet, text_encoder = load_lora_weights(unet, text_encoder, lora_weights_dir)
-        pipeline.unet = unet
-        pipeline.text_encoder = text_encoder
-    return pipeline
